tso/backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
python-dotenv
flask-cors
celery
redis
